Added changelog Started rewrite of Halo classes to be simpler and more scaleable Changed publishing mode
