+++ .github/workflows/python-publish.yml
@@ -9,9 +9,9 @@
 name: Publish to PyPi
 
 on:
-  push:
-    branches:
-      - master
+  release:
+    types: [created]
+
 
 jobs:
   build:
